[simple-storage-client] move simple-storage-client to its own crate
