+++ storage/storage-service/Cargo.toml
@@ -29,6 +29,7 @@ libra-types = { path = "../../types", version = "0.1.0" }
 libra-workspace-hack = { path = "../../common/workspace-hack", version = "0.1.0" }
 rand = { version = "0.7.3", optional = true }
 proptest = { version = "0.9.6", optional = true }
+simple-storage-client = { path = "../simple-storage-client", version = "0.1.0", optional = true }
 storage-client = { path = "../storage-client", version = "0.1.0", optional = true }
 
 [dev-dependencies]
